:recycle: [Refactor] components -> views로 변경, page 어미에 붙이기 #main
