+++ src/basic/views/BonusPointsPage.js
