+++ src/basic/views/DiscountInfoPage.js
